Mix Mastering API (#58)
* feat : file upload service

* style : reformat

* temp

* feat : mix mastering service

* feat : User entity name change : User -> Users

* feat : mixMastering service

* feat : mixmastering Service

* feat : albumart detail 유저정보 추가

* feat : user 정보 추가

* temp

* temp

* temp

* temp

* delete : local file

* delete : local file

* delete : local file

* delete : local file

* delete : local file

* delete : local file

* delete : local file

* delete : local file

* delete : local file

* delete : local file

* delete : local file

* delete : local file

* feat : 추상화 테스트

* feat : api prefix 어노테이션 추가

* refactor : 이름변경

* feat : admin service

* feat : adming service

* feat : admin service

* feat : 검색조건 추상화

* build : gitignore

* build : gitignore

* build : gitignore

* build : gitignore

* build : gradlew

* refactor : rename

* feat : albumArt 일부수정

* feat : admin albumArt

* feat : admin sales

* feat : MixMasteringService

* feat : MixMasteringService

* feat : get list 예외처리 삭제

* temp

* feat : 목록조회 수정

* feat : 목록조회 수정

* feat : 목록조회 수정

* Revert "feat : 목록조회 수정"

This reverts commit 037610b7ab570b1d9c4bcfbe8a953bce1791f500.

* temp

* feat : query 일부조건 추가

* feat : mixmastering 일부 수정

* feat : recording 일부 수정

* feat : init data 코드 분리

* feat : init data 코드 분리

* feat : init data 코드 분리

* feat : init data 코드 분리

* feat : recording 설계 변경

* temp

* temp

* feat : 검색 로직 수정

+++ src/main/java/com/study/palette/module/mixMastering/service/MixMasteringService.java
@@ -1,19 +1,21 @@
 package com.study.palette.module.mixMastering.service;
 
 
-import com.study.palette.common.PaletteUtils;
 import com.study.palette.common.dto.PaginationDto;
 import com.study.palette.common.dto.PagingDto;
 import com.study.palette.module.mixMastering.dto.CreateMixMasteringDto;
 import com.study.palette.module.mixMastering.dto.MixMasteringDetailDto;
 import com.study.palette.module.mixMastering.dto.MixMasteringDto;
+import com.study.palette.module.mixMastering.dto.MixMasteringsDto;
 import com.study.palette.module.mixMastering.dto.query.FindMixMasteringQuery;
 import com.study.palette.module.mixMastering.entity.MixMasteringInfo;
-import com.study.palette.module.mixMastering.entity.MixMasteringLicenseInfo;
+import com.study.palette.module.mixMastering.entity.MixMasteringRequest;
 import com.study.palette.module.mixMastering.exception.MixMasteringErrorCode;
 import com.study.palette.module.mixMastering.exception.MixMasteringException;
 import com.study.palette.module.mixMastering.repository.MixMasteringRepository;
+import com.study.palette.module.mixMastering.repository.MixMasteringRequestRepository;
 import com.study.palette.module.users.entity.Users;
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -30,29 +32,20 @@ import org.springframework.transaction.annotation.Transactional;
 public class MixMasteringService {
 
   private final MixMasteringRepository mixMasteringRepository;
+  private final MixMasteringRequestRepository mixMasteringRequestRepository;
   private final ModelMapper modelMapper;
 
   @Autowired
   public MixMasteringService(MixMasteringRepository mixMasteringRepository,
-      ModelMapper modelMapper) {
+      MixMasteringRequestRepository mixMasteringRequestRepository, ModelMapper modelMapper) {
     this.mixMasteringRepository = mixMasteringRepository;
+    this.mixMasteringRequestRepository = mixMasteringRequestRepository;
     this.modelMapper = modelMapper;
-  }
-
-  @Transactional(readOnly = true)
-  public PaginationDto<MixMasteringDto> getMixMasterings(Pageable pageable) {
-    Long count = mixMasteringRepository.count();
-    List<MixMasteringDto> mixMasterings = mixMasteringRepository.findAll(pageable).stream()
-        .map(MixMasteringDto::new).collect(Collectors.toList());
-
-    PaginationDto<MixMasteringDto> row = PaginationDto.of(new PagingDto(pageable, count),
-        mixMasterings);
-    return row;
   }
 
   /* MixMastering 필터 포함 조회*/
   @Transactional(readOnly = true)
-  public PaginationDto<MixMasteringDto> getixmMasterings(FindMixMasteringQuery query,
+  public PaginationDto<MixMasteringsDto> getMixMasterings(FindMixMasteringQuery query,
       Pageable pageable) {
     Long count = mixMasteringRepository.count();
 
@@ -60,25 +53,22 @@ public class MixMasteringService {
       return PaginationDto.of(new PagingDto(pageable, count), List.of());
     }
 
-    List<MixMasteringDto> artists = mixMasteringRepository.findAll(query, pageable)
+    List<MixMasteringsDto> mixmasterings = mixMasteringRepository.findAll(query, pageable)
         .stream()
-        .map(data -> modelMapper.map(data, MixMasteringDto.class))
+        .map(data -> modelMapper.map(data, MixMasteringsDto.class))
         .collect(Collectors.toList());
 
-    PaginationDto<MixMasteringDto> row = PaginationDto.of(new PagingDto(pageable, count),
-        artists);
-
-    return row;
+    return PaginationDto.of(new PagingDto(pageable, count), mixmasterings);
   }
 
+  /* MixMastering 단일 조회*/
   @Transactional(readOnly = true)
   public MixMasteringDetailDto getMixMastering(String id) {
-    Optional<MixMasteringInfo> mixMasteringInfo = mixMasteringRepository.findById(
-        UUID.fromString(id));
+    Optional<MixMasteringInfo> mixMasteringInfo = mixMasteringRepository.findById(UUID.fromString(id));
+
     if (mixMasteringInfo.isEmpty()) {
       throw new MixMasteringException(MixMasteringErrorCode.MIX_MASTERING_NOT_FOUND);
     }
-    log.info("*MixMastering* : " + mixMasteringInfo.get().getMixMasteringLicenseInfos());
 
     return new MixMasteringDetailDto(mixMasteringInfo.get());
   }
@@ -93,7 +83,7 @@ public class MixMasteringService {
 
   /* MixMastering 수정*/
   @Transactional
-  public void updateMixMastering(String id, MixMasteringDto mixMasteringUpdateReqeustDto,
+  public void updateMixMastering(String id, MixMasteringDto mixMasteringDto,
       Users user) {
     MixMasteringInfo mixMasteringInfo = mixMasteringRepository.findById(UUID.fromString(id))
         .orElseThrow(
@@ -104,31 +94,43 @@ public class MixMasteringService {
       throw new MixMasteringException(MixMasteringErrorCode.MIX_MASTERING_NOT_YOURS);
     }
 
-    PaletteUtils.myCopyProperties(mixMasteringUpdateReqeustDto, mixMasteringInfo);
-
-    // update 전 초기화
-    mixMasteringInfo.getMixMasteringLicenseInfos().clear();
-//        mixMasteringInfo.getMixMasteringFile().clear(); TODO 파일 구현 후 추가
-
-    mixMasteringUpdateReqeustDto.getMixMasteringLicenseInfos().stream().forEach(license -> {
-      mixMasteringInfo.getMixMasteringLicenseInfos()
-          .add(MixMasteringLicenseInfo.from(license, mixMasteringInfo));
-    });
-
-    //TODO 파일 구현 후 추가
-//        mixMasteringUpdateReqeustDto.getMixMasteringLicenseInfo().stream().forEach(license -> {
-//            mixMasteringInfo.getMixMasteringLicenseInfo().add(MixMasteringLicenseInfo.from(license, mixMasteringInfo));
-//        });
-
-    mixMasteringRepository.save(mixMasteringInfo);
+    mixMasteringRepository.save(mixMasteringDto.toEntity(mixMasteringInfo));
   }
 
   /* MixMastering 삭제*/
   @Transactional
-  public void deleteMixMastering(String id) {
+  public void deleteMixMastering(String id, Users user) {
     MixMasteringInfo mixMasteringInfo = mixMasteringRepository.findById(UUID.fromString(id))
         .orElseThrow(
             () -> new MixMasteringException(MixMasteringErrorCode.MIX_MASTERING_NOT_FOUND));
+
+    //본인이 작성한 글인지 체크
+    if (!mixMasteringInfo.getUsers().getId().equals(user.getId())) {
+      throw new MixMasteringException(MixMasteringErrorCode.MIX_MASTERING_NOT_YOURS);
+    }
+
     mixMasteringRepository.delete(mixMasteringInfo);
+  }
+
+  /* MixMastering 구매의뢰*/
+  @Transactional
+  public void createMixMasteringRequest(String id, Users users) {
+    MixMasteringInfo mixMasteringInfo = mixMasteringRepository.findById(UUID.fromString(id))
+        .orElseThrow(() -> new MixMasteringException(MixMasteringErrorCode.MIX_MASTERING_NOT_FOUND));
+
+    //오늘 이미 구매의뢰를 했는지 체크
+    Optional<MixMasteringRequest> mixMasteringRequest = mixMasteringRequestRepository.findByMixMasteringInfoAndUsersAndCreatedAt(
+        users, mixMasteringInfo, LocalDateTime.now());
+
+    if (mixMasteringRequest.isPresent()) {
+      throw new MixMasteringException(MixMasteringErrorCode.MIX_MASTERING_REQUEST_ALREADY_EXISTS);
+    }
+
+    mixMasteringRequestRepository.save(MixMasteringRequest.builder()
+        .mixMasteringInfo(mixMasteringInfo)
+        .users(users)
+        .createAt(LocalDateTime.now())
+        .build()
+    );
   }
 }
